[YS-559] refactor: 게시글 더보기 ui 컴포넌트 분리
